requirements.txt
aiosqlite==0.21.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.2
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
openai==1.59.5
